Refactor IPFS util common code
